Correct the detection of an existing Nix installation.
It turns out this detection is completely ineffectual for most
users, because of our elevation with sudo. We forcibly
remove Nix from the environment!

    utm@utm:~$ sudo env SHELL=/bin/bash nix
    + sudo env SHELL=/bin/bash nix
    env: ‘nix’: No such file or directory

This means that people have been able to reinstall Nix overtop Nix
for some time... I've done that at least once today? Should that
be allowed? Are they causing themselves harm? I don't know, and
am open to input on this question.

Anyway, if we do want to prohibit this we can... is this a better
test? How thorough do we want our test to be?