Add --set-home to sudo invocation
It is assumed that `dirs::home_dir()` refers to root's home directory.
This is used, for example, to setup a root profile with nix-env in
`SetupDefaultProfile`.

`dirs::home_dir()` is derived from `$HOME`, but on macOS, by default
`sudo` does not set `$HOME` to root's home unless
`--set-home` is passed.

```
> sudo bash -c 'echo $HOME'
/Users/matthew
> sudo --set-home bash -c 'echo $HOME'
/var/root
```

Pass `--set-home` to the `sudo` invocation so that `$HOME` is always
set, regardless of the system's security policy.