.buildkite: rm
As awesome as Buildkite is, it is unfortunately not useful for this
setup:

* We use Buildkite to build installer binaries
* We then use GitHub Actions to run the installer binaries on a host
without Nix already installed

We don't use Buildkite for both steps because setting up agents to
install Nix on would be a pain, as erasing everything takes a while.

This is preparation for using GitHub Actions for everything in our CI
pipeline, as it is much faster and we don't need to worry about keeping
the runners running ourselves (there have been times where our Buildkite
agents were unreachable for some reason or another).