Update and unify limits across systemd and macOS. (#1126)
* Raise the NOFILE limit to 512 * 1024 * 1024 after hitting limits talking to caches

See: https://github.com/NixOS/nix/issues/11387

* Set the default stack limit to 64M

Amazon Linux 2023 and likely CentOS / RedHat sets the default stack limit to 10M, and is generally effective at building nixpkgs. One notable exception is Nix, which typically sets it stack size to 64M. When it could not do that, a test around the stack overflow behavior failed. This was fixed in https://github.com/NixOS/nix/pull/10903 to use less stack. However, since Nix typically can use a 64M stack we set that as the default max.

We should not treat this value as magically correct. If this value causes problems, we should freely raise it. It is possible infinity is the right answer. 64M is attempting to be a "good guess" at a universally applicable good start.

For future readers, this bug report to systemd regarding LimitSTACK and DefaultLimitSTACK might be useful: https://github.com/systemd/systemd/issues/34193.

* Echo TasksMax from the systemd unit to the macOS plist

---

Also:

* Fixup the unused init parameter to ProvisionDeterminateNixd

* Fixup clippy warnings around borrows

* Delete dead code in configure_init_service