Update CONTRIBUTING.md (#1138)
Drop the note about who maintains it and why:


> an effort to explore Nix installer ideas

Outdated. We have the best installer, and it isn't an exploration anymore.

> Determinate Systems has no plans to monetize
We don't.

> or relicense `nix-installer`.
We can't, since we don't have a CLA.

> If your enterprise requires a support contact in order to adopt a tool, please contact Determinate Systems and something can be worked out.
We don't/won't do this. Companies and enterprises can become customers of Determinate Systems for FlakeHub and get all our goodies. We're not selling ad-hoc consulting services like this would suggest.