fixup: we always thought the service was started
The string "not running" contains "running", so we'd always set
`service_started = true;` here. That's wrong, so instead check for "does
not contain 'not running'".

This does open us up to thinking that anything that is not "not running"
is running, but we can address that later, if it becomes an issue.

Also I'm pretty sure we don't need this at all, since it's always "not
running" after it mounts the volume... But I'll leave that for later.