Minor macOS code fixups (#1346)
* CreateNixVolume: use constant for darwin-store service label

* BootstrapLaunchctlService: remove exit code 37 check

This is returned with the message "Operation already in progress".
However, `print` doesn't actually start anything, so I highly doubt this
can ever trigger. This has existed since the beginning of this file, but
I believe it's a relic of some leftover testing with `launchctl load` /
`launchctl unload` / `launchctl bootstrap` / `launchctl bootout`, which
likely could return this.