fix format issues with 0.27.0 merge
In the course of the 0.27.0 merge we were having trouble with the lint
workflow that ~enabled me to accidentally a few format issues in.

I introduced a format-only diff when we fiddled with a few different
ways of disabling '--determinate' features. I inadvertently swapped
the order of these when merging an item back in.

I also left an errant space in flake.nix.

Someone else left an errant one in assemble_installer.py :)