flake: use static builds on Darwin
Fix a bug where the installer would link against a `libiconv` in the
Nix store.

This would silently fail to load (falling back to the system
`libiconv`, I suppose? I’m not sure) when the Nix store was not
present – as is usually the case when running the installer – but
after installation, the store path could be populated by normal use
of Nix. If that happened, then upon uninstallation, the Nix store
`libiconv` would be loaded, the installer would delete the `/nix`
partition, and then the next time a system library indirectly caused
`dyld` to check up on what’s going on with the loaded dynamic
libraries, it would try to read from an address mapped to the library
in the now‐deleted Nix store and crash.

Now that `pkgsStatic` works after Randy’s Darwin rework to provide
mostly‐static builds (system libraries and frameworks are dynamically
linked, but libraries from Nixpkgs are linked statically), we can build
a truly standalone installer that has no dependency on `/nix/store`,
fixing the crash for good.