flake: migrate to crane
Crane handles cross‐compilation a bit better and seems to be
generally more actively maintained than Naersk.

We use Nixpkgs’ stock Rust toolchains instead of bringing in fenix,
as Nixpkgs provides everything we need and tracks upstream stable
Rust well. This also simplifies the cross‐compilation story.

This doesn’t build the documentation like the Naersk build did,
which I guess might be useful to ensure that the documentation
always builds? That could be added on as a check if desired, using
`craneLib.cargoDoc`.