Refactor diagnostics into Feedback, replace ours with detsys-ids-client
This replaces the Diagnosticdata and pervasive feature-checking with a trait with a null implementation by default.

The null implementation is more ergonomic for a couple reasons

1. needing to feature-check everywhere made it harder to write correct code for both cases.
2. Option's can't be map()'d and other combinators with async closures.

This also introduces the possibility of feature flagging behavior.
The feature flags come from the `feedback` provider, and the null version has none.

If you specify a `--diagnostic-endpoint` of `file://....` the diagnostics are sent to that path.
The format of this output has changed.
With a file endpoint, you can set the `DETSYS_IDS_CHECKIN_FILE` environment variable to a file to provide the feature flag data.

If you specify a `--diagnostic-endpoint` of an HTTP(S) address, the specific path is truncated and replaced with `/check-in` and `/events/batch`.

Note the default endpoint is now set to None, which means use the default backend.
This is because the specific host used for diagnostics is identified using an SRV record for failover and traffic shaping purposes.

Overall, this change replaces the initial and ad-hoc IDS client with a crate dedicated to the task.