ci: always validate that Release checks were successful
When "needs" jobs fail, they'll make any jobs that depend on them
be "skipped". "Skipped" jobs count as a success for the purposes of
required checks. We always want this final step to run, even if the
"needs" jobs failed (so that this "omnibus" check will halt merging if
the "needs" jobs fail).