PlaceNixConfiguration: don't write `ssl-cert-file` to nix.custom.conf if determinate and macOS
determinate-nixd will handle the `ssl-cert-file` setting in its managed
nix.conf, which will override the `ssl-cert-file` that may already be
configured in the user's nix.custom.conf.

This is _NOT TRUE_ for non-Determinate installs, since the `!include
nix.custom.conf` statement is at the _bottom_ of nix.conf on
non-Determinate installs, meaning it will override whatever is set in
nix.conf; for Determinate installs, the include statement is at the
_top_ of nix.conf, meaning the Determinate-managed config will override
anything set in the nix.custom.conf (such as `ssl-cert-file`).