Explicitly create `/nix` folder on installation
Hey everyone,
while using this installer, when using the uninstaller it won't delete the `/nix` folder. This is a nuisance on e.g. Fedora Atomic images, where you'd have to go through extra steps to delete `/`.
Checking that the installer will only revert the items it did while installing, I think that explicitly adding `/nix` to `PATHS` would solve the issue.
Thanks! Please feel free to close if it's not relevant.