Pass --speed-limit and --speed-time when passing --retry (#1637)
Some users report that the interior curl request stalls ~forever.
This patch sets a low speed-limit of 20% of modern DSL throughput.
If a user's download rate is just below the limit, it will take approximately five minutes to download.

When the speed limit trips, it will retry up to three times.

Hopefully this doesn't impact users, but if you are impacted, please open a ticket.